Elimina duplicata de AgendaMedicoView.
